requirements.txt
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.1.0
djangorestframework==3.14.0
html2text==2020.1.16
Pillow==10.1.0
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
